mullvad: Updated all available US areas
Some old files (NV, GA, MA, NC,...) are not in the lists anymore.
Leaving them in for now.
For the states, I picked the cities located in that state.

The strange thing is when I downloaded the files, UDP port was 1197.
Today it's 1196. Both seem to work.

Available now from Mullvad:
- Atlanta GA
- Chicago IL
- Dallas TX
- Denver CO
- Los Angeles CA
- Miami FL
- New York, NY
- Phoenix AZ
- Piscataway NJ
- Salt Lake City UT
- Seattle WA
- Secaucus NJ